+++ lib/base/custom_layers/models/icons.dart
@@ -1,0 +1,24 @@
+const String availabilityIcon = """
+<svg xmlns="http://www.w3.org/2000/svg" width="24.000000000000004" height="24.000000000000004" viewBox="0 0 24 24">
+      <circle fill="#64BE14" cx="12" cy="12" r="12"/>
+      <path opacity="0.1" d="M12,2c5.514,0,10,4.486,10,10s-4.486,10-10,10S2,17.514,2,12S6.486,2,12,2 M12,0C5.373,0,0,5.373,0,12   s5.373,12,12,12s12-5.373,12-12S18.627,0,12,0L12,0z" style="fill: rgb(34, 34, 34);"/>
+      <polygon fill="#FFFFFF" points="10.628,16.53 6.386,12.287 7.8,10.873 10.628,13.702 16.992,7.338 18.406,8.752 "/>
+    </svg>
+""";
+
+const String partialAvailabilityIcon = """
+<svg xmlns="http://www.w3.org/2000/svg" width="24.000000000000004" height="24.000000000000004" viewBox="0 0 24 24">
+      <circle fill="#FF9000" cx="12" cy="12" r="12"/>
+      <path opacity="0.1" d="M12,2c5.514,0,10,4.486,10,10s-4.486,10-10,10S2,17.514,2,12S6.486,2,12,2 M12,0C5.373,0,0,5.373,0,12   s5.373,12,12,12s12-5.373,12-12S18.627,0,12,0L12,0z" style="fill: rgb(34, 34, 34);"/>
+      <polygon fill="#FFFFFF" points="14,14 8,14 8,12 12,12 12,5 14,5 "/>
+    </svg>
+""";
+
+const String unavailabilityIcon = """
+<svg xmlns="http://www.w3.org/2000/svg" width="17.6" height="17.6" viewBox="0 0 24 24">
+      <circle fill="#DC0451" cx="12" cy="12" r="12"/>
+      <path opacity="0.1" d="M12,2c5.514,0,10,4.486,10,10s-4.486,10-10,10S2,17.514,2,12S6.486,2,12,2 M12,0C5.373,0,0,5.373,0,12   s5.373,12,12,12s12-5.373,12-12S18.627,0,12,0L12,0z" style="fill: rgb(34, 34, 34);"/>
+      <rect x="5.636" y="11" transform="matrix(0.7071 -0.7071 0.7071 0.7071 -4.9706 12)" fill="#FFFFFF" width="12.728" height="2"/>
+      <rect x="11" y="5.636" transform="matrix(0.7071 -0.7071 0.7071 0.7071 -4.9706 12)" fill="#FFFFFF" width="2" height="12.728"/>
+    </svg>
+""";