+++ lib/base/custom_layers/pbf_layer/charging/modal/charging_modal_models.dart
@@ -12,7 +12,7 @@ class ChargingItem {
   Set<String> capabilities = {};
   ChargingItem.fromJson(Map map)
       : uid = map["uid"]?.toString(),
-        name = map["name"]?.toString(),
+        name = (map["name"] ?? map["address"]).toString(),
         address = map["address"]?.toString() ?? "",
         city = map["city"]?.toString() ?? "",
         postalCode = map["postal_code"]?.toString() ?? "",
@@ -24,21 +24,32 @@ class ChargingItem {
     for (final detail in evses ?? []) {
       if (detail.connectors == null) return;
       for (final item in detail.connectors!) {
+        final maxElectricPower = item["max_electric_power"] is num
+            ? ((item["max_electric_power"] / 1000) as double).floor().toString()
+            : '';
         connectors[item["standard"].toString()] = ChargingConnector(
           item["standard"]?.toString() ?? '',
-          item["max_amperage"]?.toString() ?? '',
+          maxElectricPower,
         );
       }
       capabilities.addAll(detail.capabilities ?? []);
     }
   }
+
+  int get showCapacity {
+    return (evses ?? [])
+        .where(
+          (evse) => evse.status == 'UNKNOWN' || evse.status == 'STATIC',
+        )
+        .length;
+  }
 }
 
 class ChargingConnector {
   final String standard;
-  final String maxAmperage;
+  final String maxElectricPower;
 
-  ChargingConnector(this.standard, this.maxAmperage);
+  ChargingConnector(this.standard, this.maxElectricPower);
 }
 
 class ChargingDetail {