+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.7.0
+🎉 Support for Flutter 3.3.3
+🎉 Support trufi-core_v3.2.0-stadtnavi-lts
+🎉 Overlapping icons
+🎉 Improve POI's Layers
+
 ## 1.6.0
 🎉 Support for Flutter 3.3.3
 🎉 Support trufi-core_v3.2.0-stadtnavi-lts